+++ src/io/flutter/run/SdkFields.java
@@ -40,7 +40,7 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.TimeUnit;
 
 /**
- * Fields used when launching an app using the Flutter SDK (non-bazel).
+ * Fields used when launching an app using the Flutter SDK.
  */
 public class SdkFields {
   private static final @NotNull Logger LOG = PluginLogger.createLogger(SdkFields.class);