+++ src/io/flutter/run/daemon/FlutterApp.java
@@ -28,12 +28,8 @@ import com.intellij.openapi.util.text.StringUtil;
 import com.intellij.util.EventDispatcher;
 import com.intellij.util.concurrency.AppExecutorUtil;
 import com.jetbrains.lang.dart.ide.toolingDaemon.DartToolingDaemonService;
-import de.roderick.weberknecht.WebSocketException;
-import io.flutter.FlutterMessages;
 import io.flutter.FlutterUtils;
 import io.flutter.ObservatoryConnector;
-import io.flutter.bazel.Workspace;
-import io.flutter.bazel.WorkspaceCache;
 import io.flutter.dart.DtdRequest;
 import io.flutter.dart.DtdUtils;
 import io.flutter.dart.FlutterDartAnalysisServer;
@@ -48,7 +44,6 @@ import io.flutter.utils.MostlySilentColoredProcessHandler;
 import io.flutter.utils.ProcessAdapter;
 import io.flutter.utils.ProgressHelper;
 import io.flutter.utils.StreamSubscription;
-import io.flutter.utils.UrlUtils;
 import io.flutter.utils.VmServiceListenerAdapter;
 import io.flutter.vmService.ServiceExtensions;
 import io.flutter.vmService.VMServiceManager;
@@ -233,24 +228,6 @@ public class FlutterApp implements Disposable {
     LOG.info(command.toString());
 
     Consumer<String> onTextAvailable = null;
-
-    if (WorkspaceCache.getInstance(project).isBazel()) {
-      Workspace workspace = WorkspaceCache.getInstance(project).get();
-      assert (workspace != null);
-      String configWarningPrefix = workspace.getConfigWarningPrefix();
-      if (configWarningPrefix != null) {
-        onTextAvailable = text -> {
-          if (text.startsWith(configWarningPrefix)) {
-            FlutterMessages.showWarning(
-              "Configuration warning",
-              UrlUtils.generateHtmlFragmentWithHrefTags(text.substring(configWarningPrefix.length())),
-              null
-            );
-          }
-        };
-      }
-    }
-
     final ProcessHandler process = new MostlySilentColoredProcessHandler(command, onTextAvailable);
     Disposer.register(FlutterDartAnalysisServer.getInstance(project), process::destroyProcess);
 
@@ -703,7 +680,6 @@ class FlutterAppDaemonEventListener implements DaemonEvent.Listener {
   private @Nullable String appVmServiceUri;
   private static final @NotNull DtdUtils dtdUtils = new DtdUtils();
 
-
   FlutterAppDaemonEventListener(@NotNull FlutterApp app, @NotNull Project project) {
     this.app = app;
     this.progress = new ProgressHelper(project);
@@ -864,7 +840,10 @@ class FlutterAppDaemonEventListener implements DaemonEvent.Listener {
         }
       });
     }
-    catch (InterruptedException | java.util.concurrent.ExecutionException | WebSocketException e) {
+    // The Dart plugin changed its checked WebSocket exception type when it replaced
+    // Weberknecht. Catching the common base type keeps this call compatible with both APIs.
+    // We can potnetitaly catch WebSocketException in the future from Dart plugin, but since it's only used for logging, the general exception is ok
+    catch (Exception e) {
       LOG.error("Exception while sending DTD request", e);
     }
   }